improve : improve OSRMManager * move buildInstructions to OSRMManager * create osrmhelper to contain apis need it for osrmmanager
